feat(tofu): boostrap configuration for main configuration
opentofu/bootstrap contains a minimal OpenTofu configuration that creates an encrypted bucket on GCP.
This bucket is where the main configuration will store its state. The state of this bootstrap configuration
is versioned. This is safe since it doesn't contain the encryption key itself and since this should be run
often (or ever again) so there won't be any conflict.